Solution to Logical AND Operation

+++ Logical AND Operation@@67882a731b4392d8fc1577dc/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     if(a>0 && b>0){
         printf("True");
     }
